Require Meson >=1.5 for Turnip in Mesa 25.2+
Starting with Mesa 25.2, Turnip requires features introduced in
Meson 1.4 and above. On Ubuntu 24.04, the default Meson version
(1.3.2) is too old and leads to the following error during build:

  meson.build:9:18: ERROR: Meson version is 1.3.2 but project requires >= 1.4.0

This is not an issue in Mesa 25.1 and below.

Meson 1.5 (1.5.1-2ubuntu1~24.04.1) is available via apt, and is
fully backward-compatible. Builds tested and confirmed working
with meson-1.5.

Signed-off-by: Shankar Vallabhan <v3kt0r-87>